crafted HTTP request and parse methods in CreateTripController; attempted to modify JSON request URL to ask for narrower subset of information from API; initMap call returns too much data and in strange forms so parse method cannot interpret it
